Create all website files
I should have done multiple commits even though this is a small
project, but I'm not quite into the habit yet. This is a small test
project from TOP to test my ability to create a webpage of their design
using the skills I have learned with CSS thus far. This project is
primarily set up to test my knowledge of using FlexBox to arrange items
on a webpage and make them dynamic to the page size.